FEAT: Decoupled shooting method from radial solver

+++ TidalPy/RadialSolver/boundaries/surface_bc.pxd
@@ -1,6 +1,6 @@
 cdef void cf_get_surface_bc(
     double* boundary_conditions_ptr,
-    size_t* bc_model_ptr,
+    int* bc_model_ptr,
     size_t num_bcs,
     double radius_to_use,
     double bulk_density_to_use,
